New approach to responsive, mobile-only images.
This uses the <span><img></span> technique detailed by Email on Acid which uses a span to display an image on mobile-devices and hides the image within.  The initial mobile-only image implementation was handled by image itself, but moved to it's own renderer because there are a variety of important validations that made image unnecessary complex.