Removed incompatible block display from button-style links.
If a link had a mobile mode of "button" an attempt would be made to ensure the link displayed as block in the desktop version to achieve consistent spacing.  Outlook doesn't support "display: block" as a style attribute of anchor tags.  I've removed the inline style and added the display attribute to the responsive button CSS.

For the time being, the required pattern for links that turn into buttons is to wrap them in <div>s if you have multiple stacked vertically.