New Element framework to make Renderers cleaner, more maintainable.
The helper methods in the Renderer::Base framework was passing tag, attribute and style hashes around and it was getting out of hand.  The various ways elements with multiple classes were being handled was particularly unpleasant.  I've introduced a new Element class that encapsulates the tag, its classes, attributes and styles.  It handles producing the HTML and is much cleaner to pass around.