The Link renderer can now be used in the text version of the email
Presently, this allows links specified inside of the text vesion to be tagged using the same {a} format used within the html version.

Also, removed the <div>-wrapping behavior for block-style elements.  This was not implemented correctly and caused only the declaration <a> to be wrapped - e.g. <div><a></div>link text</a>.  Need to rethink how to do that.