Support link-breaks in minified CSS
The minifier support class now stores its compressor instances in the view rather than internally.  This ensures we create a new compressor with each rendering.  Both css compressor uses the context to determine if it's rendering an email and if so, applies a maximum line length to avoid problems with legacy mailers and clients that break long lines otherwise.