Added a new mobile-style tag for declaring reusable styles
In a recent commit intended on simplifing responsive email styles, the ability to name a mobile style and reuse it elsewhere was temporarily disabled.  The functionality returns with the new {mobile-style name="" style=""} tag which allows the author to declare a mobile style by name and specify its css attributes.  This declares a universal responsive class that can then be applied to any responsive element.