Fixed bug related to button-style mobile links, simplified test cases
Recent changes to the Responsive::Base class inadvertently disabled the feature that wraps button-style responsive links in <div>s for uniform layout across email clients.  This is restored by making use of the array of Responsive::Rules applied to the element - testing any for block?-status and including the <div> in such a case.

Also, disabled tag linking by default to simplify the test results.  Tag linking is enabled in specific tests but disabled for the remainder.  Unnecessary repetition.