More traditional logging format for server output
One of the problems I repeatedly had with Inkcite was that once I addressed all of the errors and warnings in my email, there was no clear indication from the server's output that I actually had a clean email.  (Meaning the errors from the previously flawed rendering would sit at the bottom of the terminal with no output following a clean rendering.)

Now it displays a timestamp'd log that first announces you've requested and the email (and what environment, format and version) that was rendered.  Any errors associated with that build are also timestamps so that it is easy to associate a given refresh with the output it produces.