Fixed source-encoding config
My original implementation expected that the configuration file would include the entire File.open-specific string (e.g. "r:from-encoding:to-encoding") to open a file read-only and then convert from a non-standard encoding to UTF-8.  An incorrectly formatted string would cause Inkcite to crash or open files as writeable.

Now, Inkcite forces the mode to read and appends the from- and to-encodings based on whether or not the configuration value is specified.