Fixed failing tests
The change to support fpo text on placeholder images broke the tests because it was specifying the text-parameter using the dimensions if an fpo message wasn't supplied.  Corrected that problem.

Also, the tests were failing because source.tsv was renamed to helpers.tsv.  The file was incorrectly named in the test's project.