Failsafes now support regex
If an includes or excludes failsafe rule starts and ends with '/' it is treated as a regular expression rather than a string.  Fixed a problem where failsafe was looking for "does-not-include" rather than "excludes" which makes mo' sense.  Some code cleanup because clarity rules and we aren't charged by the letter.