No more default font shenanigans
Removed the special-case handling for default font-family.  Instead, if a font-family is specified for a given "font" declaration, then it is used, regardless of whether or not it is the default.  This ensures that when appropriate, the font-family can be reiterated in every td, div or span.  This is necessary to fix problems in third-party engines with Inkcite-produced emails.