Less duplication on background-color
For table-based elements (e.g. {td}) it was setting both bgcolor attribute and background-color styles to the same value.  This appears to be unnecessary duplication.  Email clients read the bgcolor attribute reliably.  Now, background-color is only populated when mandated by CSS such as a background image being present or when overriding on mobile.