Footnote support
Particularly when there are multiple versions of an email, footnotes can get out-of-whack and hard to manage.  Now, Inkcite will manage your footnotes automatically based on the order the footnotes appear in the email.

Use the {footnote text=""} tag to place a footnote reference inside of the email body content.  Inkcite will assign a unique numeric symbol if the symbol attribute is omitted.  Reuse a footnote by assigning an ID the first time it is declared and then reference it in {footnote} tags further down.

Include the {footnotes} tag in your email when you want the previously-specified footnote text references to be injected in order.