Support for version-specific helper overrides, move helper loading to the View
This change simplifies applying different styles, colors and formatting to a multi-version email. Previously the designer was limited to a single helper.tsv file.  This change allows the designer to override helper declarations on a per-version basis.  Moved the helper loading logic to the View (where, in hindsight, it should have been originally).  Added logic to look for a helper-(version).tsv file and load that after the original helper file.

Updated the tests to reflect properties moving from Email to View.