Fixed a problem where {footnotes} could produce run-on markup
When rendering the email version of the project, breakup multiple footnotes with a carriage return.  This is done regardless of whether or not the built-in or a custom template is used.  This minimizes the possibility of a problem where run-on markup could be created in emails that have a large number of footnotes.