Fixed an exception thrown if a source file is missing
When cloning an existing project, if one of the expected source files doesn't exist, an exception would be thrown.  Now it tests for the existance of the file and only copies those that originate from the cloning source.  This problem was detected when I tried to clone a directory that didn't have a source.txt version of said email.