Use the tag stack for custom helpers, keeping closing tags DRY
It's bugged me for a long time that custom helpers (e.g. {feature ...}) didn't use the tag stack.  If you had a variable used in both the open and closing HTML, you had to repeat it in the closing tag (e.g. {feature name=value} and {/feature name=value}).  Y'all know how I feel about repeating yourself in HTML.  Here there be dragons.

So now the custom Property (a.k.a. custom helper) renderer determines if the tag is an open or close tag and uses the tag stack accordingly.