New Increment {++} Helper
There is a very explicit use case where this Helper can be used.  It allows the designer to request a unique numeric counter associated with an ID, similar to how the auto-incrementing link IDs work.  The use case for this Helper is: {++ id="my-counter"} which will produce 1, 2, 3 ... n, depending on the number of times it appears in the email.  Useful for advanced link tagging and editable field integration in Siebel.