Adding footnotes.tsv support
This adds support for a central footnotes.tsv file to declare footnotes and copy them from project to project.  This is particularly useful when working in multi-version emails where footnotes are used across several partials with different orders per version.

In View, added a method exposing the footnotes_tsv file which is parsed on demand.  Added a constant defining the name of the file.

The Footnote renderer adds an :active attribute used to track footnotes loaded from tsv, but not yet referenced in HTML.  Changed logic to assign symbols to footnotes that are declared but inactive.