Improvements to Litmus analytics - auto-request report ID, dynamic tracking code HTML
Including the {litmus} tag automatically requests a new report ID if one is not provided. The Litmus-assigned ID is stored in the email version's meta data and used for subsequent builds. The Litmus API is queried for up-to-date "HTML bug" code with each build.

Had to rename Inkcite::Renderer::Litmus to Inkcite::Renderer::LitmusAnalytics to avoid naming conflicts with the Litmus API