Improved table, td background mixology
On mobile devices, background images were getting their size style attributes jumbled.  In certain circumstances it was possible for the responsive-applied background-size to be overruled by the background composite style required to support backgrounds in Gmail (desktop).  This new way of handling backgrounds is cleaner, less hacky.

Inkcite still tries to use bgcolor= attributes as a 100% safe cross-client solution for background colors.  The background style attribute is only used when images are involved.  It generates the complete background style for both desktop and mobile and then compares them to determine if a mobile-only style is needed, and whether or not !important is required.  The new composite background CSS style includes background size embedded per the CSS spec.

Updated tests to reflect the addition of the background size in test results.