Display optimized images in server mode
When running "inkcite server" the user will now see the optimized/compressed images if image optimization is enabled.  Server triggers image optimization prior to rendering the email.  Updated the email to return the optimized images path when in development.

Changed the cache path for optimized images from a hidden .images directory into images-optim to make it clear to the user what is happening and to easily compare results.

Fixed a problem where the image_optim.yml configuration file could be modified but the images in the cache wouldn't be regenerated.