Make exposed <body> visually distinct
Email clients trim off <body> statements so Inkcite's previous behavior of setting the HTML <body> color to a #background helper was undesirable.  Under certain conditions, a designer might not realize the body was exposed or un-wrapped.

Now in development mode it injects a pleasing subtle blueprint pattern ensuring that the designer knows the body is inadvertently exposed.