Fixed a problem where Footnotes could get duplicated
Under certain conditions (basically anytime a footnotes.tsv came after an inline footnote) when a footnotes.tsv file is combined with arbitrary footnotes in an email, it was possible for the IDs to get duplicated.  Now the Footnote helper collects active, numeric footnotes and then grabs the max value.  Added an active? alias for readability.