Beautify Inkcite's preview server output in development mode
Inkcite's Server command now runs the email HTML through htmlbeautifier when the environment is set to development.  Having beautified code makes it easier to diagnose problems while inspecting the source in the browser.

Added a dependency on the htmlbeautifier gem.  Updated Server to pass the HTML through the beautify() command if viewing the email in development mode.