A variety of small updates
Allow mobile-breakpoint to be defined in the project to control the point at which the email switches to responsive view.  Inherits from width if not defined.  Defaults to 480.

Use images/ if an image-host is not defined in the config.yml.  This ensures that "inkcite build" will function regardless of the status of the config.yml.  Also, disable cache-busting in production always.

Fix to enable responsive media queries on Windows phones.
Fix to remove title from the production email HTML.