The Uploader uses the email version, rather than browser, when uploading images
The Uploader now treats the browser version of the email as the one-off that it really is.  Most ESPs will host the browser version of the email on their own servers, built dynamically from the email itself.  The case where a browser.html is built is a one-off.  So the images for the email version are uploaded and then, if the browser-version is enabled in the config file, it produces said version and pushes it to the server.