Alignment-related fixes for fluid-drop tables
When fluid-drop is specified, the designer gains the most flexibility in how the floating columns vertically align if the outer table uses valign=middle by default.  So, changed the default from top to middle, which is also a more natural default since it matches how a normal table would respond.

In the td helper, removed an unread variable.  Added support for mobile=hide on fluidly-columns, meaning one of the interior columns can be hidden in the responsive view.  Lastly, updated the text to align left by default, which also better mirrors the behavior of a regular table column.

Updated the minitest to reflect the changes in the resulting HTML.