Remove font-size warning on fluid-drop TDs
I originally thought it would be helpful to warn the user when they have a fluid-drop TD without a font-size -- but after practical use, it is common to not have text directly inside of such a TD, typically having it inside of a deeper nested element.  So the warning is getting in the way of working so I am removing it.