Fixed a bug where updated images wouldn't trigger a reload, also fixed condition for beautification.
The programmatic Guard initialization was incorrectly presuming that the image paths were relative.  Guard tests against full path names so we can't watch() a string that starts with images, it needs to include that somewhere within.  Also, watching for GIF, JPG and PNG files explicitly.

Also, the HtmlBeautifier should only be invoked if minification is disabled.  This allows the dev to view the minified output in-browser if they so choose.