Preview and test specific versions of an email, plus better error messages
Inkcite's preview and test commands now accept the --version (or -v) option allowing a specific version of the email to be sent to a recipient list or to the testing service.  Updated Mailer quite considerably - now its functions accept an opts parameter and it looks for the :version attribute or sends all of them.  Exposed the send() method to simplify the implementation of test.  Refactored Mailgun and SMTP handling into an abstract base class and two concrete classes.

Inkcite test also now optionally accepts an array of test addresses - convenient for users without Litmus or Email on Acid if they want to send one or more copies of the email to different test accounts.

Better error messages in a couple of cases using ActiveSupport's strip_heredoc.