Added Litmus Scope integration - easily share your emails
Add scope method to Cli::base, with version option.  Proxies the request to the new Scope class.  Scope uses the Mail gem to assemble the Mail code, then uses https to post the email to the Litmus Scope service.  This is a first pass, so there is probably a fair bit of error checking I need to add.