Fixed a potential nil exception
When an unopened custom Helper tag was closed (e.g. {h3}...{/h2}) it was possible for an exception to be thrown in the property renderer.  The problem was caused by the code trying to access the opening opts hash when the tag stack didn't have one to provide.  Now the code verifies it has an open stack before trying to merge the parameters.