Lorem Ipsum now supports word-limited placeholder text, also the 'force' attribute
Lorem Ipsum can now have a type of 'words' with a limit allowing specific word-count placeholder text to be generated.

When the boolean 'force' attribute is supplied, it will no longer trigger an error in the email to include Lorem Ipsum.  This is useful in the rare instances where you want to include Lorem Ipsum in an email intentionally.

Also