New {redacted} tag is an alternative to Lorem Ipsum
This new tag converts the contents of its 'text' attribute into X and x characters, while preserving the symbols - e.g. Test becomes Xxxx.  The 'with' attribute allows the replacement character (which defaults to X) to be specified.  Unless 'force' is specified, an error is recorded so an email doesn't ship with redacted text within it.

This is more predicatable than random Lorem Ipsum.  It is useful for content that isn't finalized - e.g. pricing or availability dates where you want to prototype the email but need a reminder to get accurate numbers before it ships.