Allow shared/library helper.tsv files to be specified in config.yml
Updated the load_helpers method to check the project config.yml for additional files specified in the 'helpers' entry.  Treats it always as an array, supporting both a single and multiple inheritance.  Also, changed the way missing helpers are treated - instead of throwing an error, it complains at the command line.