Show animated snow in browser, development modes
There is no reason to limit the snow effect to webkit-only browsers during development.  Similarly, if a hosted browser version is being rendered, it too can include the snow effect.  This is accomplished using a new method that renders either just the -webkit prefix or both it and the non-prefixed version (which I have been assured works in Firefox, Opera and unexpectedly IE).