Adding support for images, rotation on snowflakes
If the src attribute is specified, the snowflake generator will apply it as a background image, sized to fill the snowflake's dimensions.  Adding an image disables the background color and the border-radius.

If an image is provided and boolean rotate (or rotation, for convenience) attribute is specified, a random rotation transform is applied to the flake's ending position.  Modified the with_browser_prefix method to accept a separator - so either multi-line (default) or inline (spaces) can be requested from this method.