More flexible {button} configuration
The {button} Helper made a variety of assumptions about their default appearance - drawing initial values from the #link color, primarily - and presuming that text should automatically contrast.  With the rise in popularity of the so-called Ghost button, these "smart" defaults made it challenging to code a transparent, bordered button in Inkcite.  So I'm changing the {button} so that it no longer relies on guessing and instead uses the values within helper.tsv to entirely control its appearance.