Images with alt text no longer default to 'small' font
Previously if alt text was specified on an image, Inkcite would automatically try to apply the 'small' font style to the image.  This lead to problems where image font size, line height, etc. would have to be frequenly overridden.  So now, font and its corresponding attributes have to be set explicitly which provides more control for the designer.