Link Helper HREFs now automatically tested for validity
The URL provided to the Link Helper is now tested for validity using URI's parser.  This catches problems like the time that there was a line-break in an HREF unexpectedly.  If a false-negatives is detected (e.g. URLs containing ISP substitution flags) the Link Helper supports the new 'force' boolean attribute that allows the validation to be skipped.