Strip HTML markup from image alt text
Image alt tags are suppose to be text, not markup. But it is very common to define custom Helpers that accept a title or headline that appears both as text and as image alt text - e.g.:

{img src=something.jpg alt="$title$" ...}
{h2}$title${/h2}

But if the $title$ parameter contains markup (e.g. <i>Magazine Title</i>'s Top Ten of 2016) the HTML would be displayed improperly when images were disabled.

With this change (and corresponding test) any HTML within the alt text of an {img} Helper is stripped automatically.