Two fixes related to multi-line alt text
First, the additional space (injected to fix the problem with Gmail ignoring line breaks) was only being applied when a virtual line break (literal "\n") was included in the alt text.  Now all line breaks within alt text are prefixed with a space.

Second, the original implementation put the space after the line break which resulted in strangely indented alt text on lines 2 and beyond.  Now the space is appended before the line break to have the last impact on the display when images are disabled.