Fresh thinking on {mobile-img}
The {mobile-img} tag is no longer a container.  Basically this is an ease-of-use improvement.  It was less common to have an image on desktop in the exact same spot that a different mobile image would be needed that couldn't be addressed by {img}'s mobile-src attribute.  Besides, it's easy enough to simply add mobile=hide to the {img} and get the same behavior.

So basically, it was a lot of unnecessary typing to create an empty stand-alone mobile-img - e.g. {mobile-img src=...}{/mobile-img}.

Instead, mobile-img just works now and doesn't require a closing tag.