Switched to HTML5 doctype, fixed a problem with automatic date link breaking
The previous method for breaking automatic linking of months was flawed in that trying to do a non-breaking space between month and day would get ignored in certain email clients.  Now a zero-width non-breaking space is injected in the middle of the month name ensuring the date doesn't convert to a link on iOS while maintaining a single word that can be non-breaking space'd to other values.