Instead of minifying images before rendering the HTML, minify images inline
If image minification is enabled, rather than minify the images before the HTML can be rendered, minify the images as they are requested by the local browser.  Added an OptimizedImage handler that intercepts requests for images-optim/ files and passes the file to the minifier, then lets the super class serve it up.