Added insets support to SpecialEffect, equal_distribution() takes the range, Fireworks bug fixes
Insets allow for percentage-based positions within a special effect to have a margin or buffer from the edges.  Hue range is specifiable on the fireworks effect.  Fixed a problem where positions were duplicated when multiple Fireworks effects were specified.  Support for rainbow boolean that causes an equal distribution of hues based on the number of fireworks.