New Topic List Helper
The new {topic} and {topic-list} Helpers remove the burden of building version-specific preheader article lists.  Define a {topic name="..."} and include the {topic-list} anywhere in the email.  Topics will be listed, comma-separated, in the order they are encountered during rendering.  An optional priority attribute on {topic} allows you to control the article order (higher priority first).

Added a concept of a post-processor to View allowing Helpers and other logic to install logic that gets called when all other rendering has been cmpleted but before failsafes are evaluated.